+++ src/main/java/org/sophy/sophy/controller/PlaceController.java
@@ -2,7 +2,7 @@ package org.sophy.sophy.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.sophy.sophy.common.dto.ApiResponseDto;
-import org.sophy.sophy.controller.dto.request.CityRequestDto;
+import org.sophy.sophy.domain.dto.CityRequestDto;
 import org.sophy.sophy.controller.dto.response.PlaceResponseDto;
 import org.sophy.sophy.exception.SuccessStatus;
 import org.sophy.sophy.service.PlaceService;
