+++ src/main/java/org/sophy/sophy/domain/dto/MyBookDto.java
@@ -3,7 +3,7 @@ package org.sophy.sophy.domain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import org.sophy.sophy.domain.BookCategory;
+import org.sophy.sophy.domain.enumerate.BookCategory;
 
 @Getter
 @AllArgsConstructor
