+++ src/main/java/org/sophy/sophy/domain/dto/MyPageBooktalkDto.java
@@ -3,7 +3,7 @@ package org.sophy.sophy.domain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import org.sophy.sophy.domain.BooktalkStatus;
+import org.sophy.sophy.domain.enumerate.BooktalkStatus;
 
 import java.time.LocalDateTime;
