+++ src/main/java/org/sophy/sophy/domain/enumerate/Authority.java
@@ -1,4 +1,4 @@
-package org.sophy.sophy.domain;
+package org.sophy.sophy.domain.enumerate;
 
 public enum Authority {
     ROLE_USER, ROLE_ADMIN
