+++ src/main/java/org/sophy/sophy/domain/enumerate/BookCategory.java
@@ -1,4 +1,4 @@
-package org.sophy.sophy.domain;
+package org.sophy.sophy.domain.enumerate;
 
 public enum BookCategory {
     //인문/문학/예술/과학/육아/일상_취미/어린이/청소년/IT_컴퓨터/자기계발/건강_요리/여행
