+++ src/main/java/org/sophy/sophy/infrastructure/PlaceRepository.java
@@ -1,6 +1,6 @@
 package org.sophy.sophy.infrastructure;
 
-import org.sophy.sophy.domain.City;
+import org.sophy.sophy.domain.enumerate.City;
 import org.sophy.sophy.domain.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
